refactoring FormModal
