+++ .meteor/release
@@ -1,1 +1,1 @@
-METEOR@1.7.0.3
+METEOR@1.8.0.2
