+++ .meteor/versions
@@ -1,16 +1,16 @@
-accounts-base@1.4.2
+accounts-base@1.4.3
 accounts-password@1.5.1
 allow-deny@1.1.0
-autoupdate@1.4.1
-babel-compiler@7.1.1
-babel-runtime@1.2.2
+autoupdate@1.5.0
+babel-compiler@7.2.4
+babel-runtime@1.3.0
 base64@1.0.11
-binary-heap@1.0.10
-blaze@2.3.2
+binary-heap@1.0.11
+blaze@2.3.3
 blaze-html-templates@1.1.2
 blaze-tools@1.0.10
-boilerplate-generator@1.5.0
-caching-compiler@1.1.12
+boilerplate-generator@1.6.0
+caching-compiler@1.2.1
 caching-html-compiler@1.1.3
 callback-hook@1.1.0
 check@1.3.1
@@ -20,22 +20,23 @@ ddp-common@1.4.0
 ddp-rate-limiter@1.0.7
 ddp-server@2.2.0
 deps@1.0.12
-diff-sequence@1.1.0
-dynamic-import@0.4.1
-ecmascript@0.11.1
+diff-sequence@1.1.1
+dynamic-import@0.5.1
+ecmascript@0.12.4
 ecmascript-runtime@0.7.0
-ecmascript-runtime-client@0.7.1
-ecmascript-runtime-server@0.7.0
+ecmascript-runtime-client@0.8.0
+ecmascript-runtime-server@0.7.1
 ejson@1.1.0
 email@1.2.3
 es5-shim@4.8.0
-fourseven:scss@4.9.0
+fetch@0.1.0
+fourseven:scss@4.10.0
 geojson-utils@1.0.10
 hot-code-push@1.0.4
 html-tools@1.0.11
 htmljs@1.0.11
-http@1.4.1
 id-map@1.1.0
+inter-process-messaging@0.1.0
 johnantoni:meteor-normalize@0.0.1
 jquery@1.11.11
 launch-screen@1.1.1
@@ -44,40 +45,41 @@ localstorage@1.2.0
 logging@1.1.20
 meteor@1.9.2
 meteor-base@1.4.0
-minifier-css@1.3.1
-minifier-js@2.3.5
-minimongo@1.4.4
+minifier-css@1.4.1
+minifier-js@2.4.0
+minimongo@1.4.5
 mobile-experience@1.0.5
 mobile-status-bar@1.0.14
-modern-browsers@0.1.2
-modules@0.12.2
-modules-runtime@0.10.2
-mongo@1.5.1
+modern-browsers@0.1.3
+modules@0.13.0
+modules-runtime@0.10.3
+mongo@1.6.0
+mongo-decimal@0.1.0
 mongo-dev-server@1.1.0
 mongo-id@1.0.7
 npm-bcrypt@0.9.3
-npm-mongo@3.0.7
+npm-mongo@3.1.1
 observe-sequence@1.0.16
 ordered-dict@1.1.0
-promise@0.11.1
+promise@0.11.2
 random@1.1.0
 rate-limit@1.0.9
 react-meteor-data@0.2.16
-reactive-dict@1.2.0
+reactive-dict@1.2.1
 reactive-var@1.0.11
 reload@1.2.0
 retry@1.1.0
-routepolicy@1.0.13
+routepolicy@1.1.0
 service-configuration@1.0.11
-session@1.1.7
+session@1.2.0
 sha@1.0.9
-shell-server@0.3.1
+shell-server@0.4.0
 socket-stream-client@0.2.2
 spacebars@1.0.15
 spacebars-compiler@1.1.3
-srp@1.0.10
-standard-minifier-css@1.4.1
-standard-minifier-js@2.3.4
+srp@1.0.12
+standard-minifier-css@1.5.2
+standard-minifier-js@2.4.0
 templating@1.3.2
 templating-compiler@1.3.3
 templating-runtime@1.3.2
@@ -86,6 +88,5 @@ tmeasday:check-npm-versions@0.3.2
 tracker@1.2.0
 ui@1.0.13
 underscore@1.0.10
-url@1.2.0
-webapp@1.6.2
+webapp@1.7.2
 webapp-hashing@1.0.9
