+++ imports/client/styles/_pages.scss
@@ -666,6 +666,10 @@ a.compose-mail {
   vertical-align: middle;
 }
 
+.project-actions > div {
+  display: inline;
+}
+
 dd.project-people {
   text-align: left;
   margin-top: 5px;
