+++ imports/ui/dashboard/DashboardPage.js
@@ -4,28 +4,30 @@ import { Meteor } from 'meteor/meteor';
 import { withTracker } from 'meteor/react-meteor-data';
 import { Classrooms } from '../../api/classrooms';
 import { Teachers } from '../../api/teachers';
+import { Toolbar } from '../common/Toolbar';
+import { DashboardTeachersNotAssigned } from './DashboardTeachersNotAssigned'
+import { DashboardClassrooms } from './DashboardClassrooms'
 
 export class DashboardPage extends React.Component {
   render() {
-    const { classrooms, teachers } = this.props;
+    const { classrooms, teachers, title, buttonText } = this.props;
     return (
-      <div className="page">
-        {classrooms.map(classroom => (
-          <div key={classroom._id}>
-            <h4>{classroom.name}</h4>
-            <h5>{classroom.description}</h5>
-            <p>Ratio: {classroom.ratio}</p>
-          </div>
-        ))
-        }
-        <div>
-          <h4>Unassigned teachers</h4>
-          {teachers.filter(teacher => !teacher.classroom_id).map(teacher => (
-            <div key={teacher._id}>
-              <p>{`${teacher.firstName} ${teacher.lastName}`}</p>
-            </div>
-          ))}
-        </div>
+      <div key="homeView">
+
+        <Toolbar
+          title={"Dashboard"}
+          breadcrumb={['']}
+          button={this.openModal}
+          buttonText={buttonText} />
+
+        <DashboardTeachersNotAssigned
+          {...this.props}
+        />
+
+        <DashboardClassrooms
+          {...this.props}
+        />
+
       </div>
     );
   }
