+++ package.json
@@ -6,7 +6,7 @@
     "test": "jest --config jest.config.json"
   },
   "dependencies": {
-    "@babel/runtime": "^7.0.0-beta.54",
+    "@babel/runtime": "^7.2.0",
     "bcrypt": "^3.0.0",
     "final-form": "^4.8.3",
     "history": "^4.7.2",
@@ -15,13 +15,16 @@
     "prop-types": "^15.6.1",
     "react": "^16.3.2",
     "react-bootstrap": "^0.32.1",
+    "react-dnd": "^5.0.0",
     "react-dom": "^16.3.2",
     "react-final-form": "^3.6.5",
     "react-modal": "^3.5.1",
+    "react-redux": "^5.0.7",
     "react-router-dom": "^4.2.2",
     "react-table": "^6.8.6",
     "react-timestamp": "^4.4.0",
-    "simpl-schema": "^1.5.0"
+    "simpl-schema": "^1.5.0",
+    "sortablejs": "^1.7.0"
   },
   "devDependencies": {
     "babel-preset-env": "^1.7.0",
